add frontend structure
