+++ Project/frontend/src/App.js
@@ -2,13 +2,13 @@ import React from "react";
 import './App.css';
 import {BrowserRouter as Router, Switch, Route} from "react-router-dom";
 import Nav from "./components_two/nav";
-import Leagues from "./components_two/leagues";
-import LeagueDetail from "./components_two/leagueDetail";
-import Teams from "./components_two/teams";
-import TeamDetail from "./components_two/teamDetail";
-import Players from "./components_two/players";
-import PlayerDetail from "./components_two/playerDetail";
-import News from "./components_two/news";
+import Leagues from "./components_two/leagues/leagues";
+import LeagueDetail from "./components_two/leagues/leagueDetail";
+import Teams from "./components_two/teams/teams";
+import TeamDetail from "./components_two/teams/teamDetail";
+import Players from "./components_two/players/players";
+import PlayerDetail from "./components_two/players/playerDetail";
+import News from "./components_two/general/news";
 import ListOfStadiums from "./components_two/stadium";
 import AddCountry from "./components_two/general/countries";
 import AddCity from "./components_two/general/cities";
