+++ Project/frontend/src/components_two/general/cities.js
@@ -34,7 +34,7 @@ const AddCity= () => {
         <div className="container">
             <div className="container">
       <div className="w-75 mx-auto shadow p-5">
-        <h2 className="text-center mb-4">Add A Student</h2>
+        <h2 className="text-center mb-4">add a city</h2>
         
         <div className="form-group">
           </div>
