+++ Project/frontend/src/components_two/general/countries.js
@@ -34,7 +34,7 @@ const AddCountry= () => {
         <div className="container">
             <div className="container">
       <div className="w-75 mx-auto shadow p-5">
-        <h2 className="text-center mb-4">Add A Student</h2>
+        <h2 className="text-center mb-4">add a country</h2>
         
         <div className="form-group">
           </div>
