+++ Project/frontend/src/components_two/leagues/leagues.js
@@ -1,6 +1,6 @@
 import React, {useState, useEffect} from 'react';
 import axios from 'axios';
-import './Model.css';
+//import '../Model.css';
 import {Link} from 'react-router-dom';
 function Leagues() {
