+++ Project/frontend/src/components_two/players/playerDetail.js
