+++ Project/frontend/src/components_two/teams/teamDetail.js
