+++ Project/frontend/src/components_two/teams/teams.js
@@ -1,29 +1,30 @@
 import React, {useState, useEffect} from 'react';
 import axios from 'axios';
-import './Model.css';
+import '../Model.css';
 import {Link} from 'react-router-dom';
-function Players() {
+function Teams() {
 
-    const [players, setPlayers] = useState([]);
+    const [teams, setTeams] = useState([]);
 
     useEffect( () => {
         axios({
             method:"GET",
-            url:"http://127.0.0.1:8000/api/players/",
+            url:"http://127.0.0.1:8000/api/teams/",
             mode: "no-cors"
         }).then(response => {
-            setPlayers(response.data)
+            setTeams(response.data)
         })
     },[])
     return(
         <div>
-            <h1>Players page</h1>
-            <h2>list of players</h2>
+            <h1>Teams page</h1>
+            <h2>country</h2>
+            <hr/>      
                 <hr/>      
-                {players.map(item => (
+                {teams.map(item => (
                     <div>
                     <h1 key={item.id}>
-                        <Link to={{ pathname: `/player/${item.id}/`, fromDashboard: false}}>
+                        <Link to={{ pathname: `/teams/${item.id}/`, fromDashboard: false}}>
                             {item.name}<br/>
                         </Link>
                     </h1>
@@ -32,4 +33,4 @@ function Players() {
                 </div>)
                 }
 
-export default Players;
+export default Teams;
