+++ Project/frontend/src/components_two/users/login.js
@@ -30,7 +30,7 @@ const LogIn = () => {
         <div className="container">
             <div className="container">
       <div className="w-75 mx-auto shadow p-5">
-        <h2 className="text-center mb-4">Add A Student</h2>
+        <h2 className="text-center mb-4">login</h2>
         
         <div className="form-group">
           </div>
