+++ Project/frontend/src/components_two/users/registration.js
@@ -32,7 +32,7 @@ const Registration = () => {
         <div className="container">
             <div className="container">
       <div className="w-75 mx-auto shadow p-5">
-        <h2 className="text-center mb-4">Add A Student</h2>
+        <h2 className="text-center mb-4">registration</h2>
         
         <div className="form-group">
           </div>
